No longer recommend the MockHandler in README
The MockHandler is an advanced topic that should be discussed on its
own.

Also, having it in the configuration example that people may copy/paste
is really not practical because they will have a client that doesn't
actually do anything, which is confusing to people who don't know what a
MockHandler is.

Lastly, to use the MockHandler effectively, one should be able to prime
it before it is added to a Client, but that is not possible using this
bundle, so let's not recommend doing that at all.