README clean up (#282)
* Link to Symfony from README

* Redefine prerequisites instead of requirements

The requirements as they were might suggest that people need to install
guzzle manually, but that's not the case. So that was removed from the
requirements, and requirements was renamed to prerequisites, as that
sounds less harsh and is used more often in this kind of context.

* Update installation section of README

Describe installation with Symfony Flex as default, but also offer ways
to help projects that do not use Symfony Flex.

Also stop recommending people to use dev-master, that should really not
be done unless you really know what you're doing. And people who know
what they're doing don't need documentation on how to require
dev-master.

* Introduce configuration section

How to use the bundle is different from how to configure it.
Configuration deserves its own section for discoverability.

* No longer recommend the MockHandler in README

The MockHandler is an advanced topic that should be discussed on its
own.

Also, having it in the configuration example that people may copy/paste
is really not practical because they will have a client that doesn't
actually do anything, which is confusing to people who don't know what a
MockHandler is.

Lastly, to use the MockHandler effectively, one should be able to prime
it before it is added to a Client, but that is not possible using this
bundle, so let's not recommend doing that at all.

* Remove outdated comment from configuration section in README

* Replace usage section with manual service wiring in README

Getting public services from the DI container has been discouraged by
Symfony for a while now, let's not show people how it's done, but show
them the recommended Symfony way instead, also providing a link for
those who prefer autowiring over manual wiring.

* Update plugin installation section with update Kernel code

* Expand README section on events

Also replace XML with YAML, which is used in the rest of the README as
well

* Assorted fixes to README

* Clarify which Symfony versions are supported in README

* Fix autowiring link in README

* Fix styling of event names in README

* Make clear that service aliases are not useful when autowiring in README

* Show Symfony as correct link in README